az_functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure.cosmos
bson
datetime
requests
openai
azure-identity
azure-mgmt-web
msal